Use inline HTML for heading tags, not Gitiles MD
This syntax for creating tags for linking to headings with complex names
isn't supported in CommonMark, which is the standard used by MD lint.
It seems to be valid syntax in Gitiles MD, which is why the links work.

Upstream feature request to add support for this type of section header:
DavidAnsonFOO/markdownlint#570

This is required for submariner-ioFOO/submariner-website#827.

Signed-off-by: Daniel Farrell <dfarrell@redhat.com>